Moved Wrapper class to utils, dirty solution
